third_party/blink/web_tests/fast/events/event-listener-on-attribute-inside-shadow-dom-expected.txt

CONSOLE WARNING: line 12: Element.createShadowRoot is deprecated and will be removed in M80, around February 2020. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
This tests ensures that an event listener on an attribute node inside a shadow DOM is properly unregistered when parent element of the attribute is moved to a new document.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS internals.touchStartOrMoveEventHandlerCount(document) is 1
PASS internals.touchStartOrMoveEventHandlerCount(document) is 0
PASS successfullyParsed is true

TEST COMPLETE
